+++ app/src/components/projectSettings/ProjectUserTable.tsx
@@ -7,7 +7,7 @@ import { useHandledAsyncCallback, useSelectedProject } from "~/utils/hooks";
 import { RemoveProjectUserDialog, type ProjectUser } from "./RemoveProjectUserDialog";
 import { api } from "~/utils/api";
 import { maybeReportError } from "~/utils/errorHandling/maybeReportError";
-import { useAccessCheck } from "../AccessControl";
+import { useAccessCheck } from "../AccessCheck";
 
 const ProjectUserTable = () => {
   const selectedProject = useSelectedProject().data;
